requirements.txt
ccxt
binance-connector